random-expression: fix the parsing of negative weights
The parsing of parameters use Boost.Tokenizer with
escaped_list_separator which is meant for CSV files, so it needs three
characters: the escape (defaults to `\`), the separator (`,`), and the
string delimiter (`"`).  In our case we do not want `\` to escape
anything, as we want to be able to use it to denote left-division, so
it was chosen to use `-` instead.  That was a bad idea, as
`w="min=-10"` was treated as if we had written `\10`, which makes no
sense.

That's again a sign that we should rewrite this parser.  See #131.

* vcsn/algos/random-expression.hh: Use `####` as escape character
(which is probably the same as `#`...).
* tests/python/factory.py: Exercize this case.