ipynbdoctest: check all the pages
Fixes #42.

For some reason, automaton.eliminate_state fails on some temporary
ArchLinux agents on our BF, but works properly on Prague, our
permanent ArchLinux slave.  So let's skip this.  I agree it would be
better to know exactly what is wrong, but it it really difficult to be
able to spawn a tmp slave and trying to log on it.

* tests/bin/ipynbdoctest.py: Skip tests that fail to find ipywidgets.Label.

We finally managed to fix all the errors.

* doc/notebooks/local.mk: Check all the notebooks.