eppstein: x10 in perf, do not pre-compute the sidetrack costs
The algorithm used to be a lot slower when pre-computing the sidetrack
costs. Which is resolved by an on-the-fly calculation.
The overhead in computing each costs before the algorithm is caused by
the fact that we won't have to visit all of the automaton's transitions
as most of them won't be close to the actual K shortest paths. We then
avoid a lot of useless computations.

(1) Before
(2) After

(1)    (2)
42.50  2.13 a.lightest(5, "eppstein") # a = std([a-e]?{150}), c = [a-e] -> Nmin, 500x