lightest-automaton: simplifications
* vcsn/algos/lightest-automaton.hh: Improve error message.
Reorder the if in a more natural order.
Don't check for null_transitions in paths, this cannot happen
(and does not happen in the test suite).
And use the copier's feature to copy sets of transitions.

* tests/python/lightest-automaton.py: Reorder.
Check error messages.
(check): Simplify (there is no need to treat \z in a special way), and
fix (the weight applies to the whole expression, not just the first
term).