build: avoid warnings
* vcsn/core/rat/dot.hh: Make the instantiation of visit_tuple depend
on a template parameter to avoid instantiation when not used.  This
could be see because we had warnings about it even in lal, b, which,
of course, does not instantiate `visit_tuple`.
See also http://stackoverflow.com/questions/40385953.
But also address the warning that we see with tuples: initialize
name_.