expressions: use static-if to handle tuple nodes
So far, we were using a (partial) specialization of class templates to
dispatch on whether are multitape or not.  Use static-if instead.
This is still not very elegant, but it seems better than what we had.

* vcsn/core/rat/copy.hh, vcsn/core/rat/hash.hh, vcsn/core/rat/info.hh,
* vcsn/core/rat/less.hh, vcsn/core/rat/partial-identity.hh,
* vcsn/core/rat/printer.hh vcsn/core/rat/project.hh,
* vcsn/core/rat/size.hh, vcsn/core/rat/transpose.hh:
Use static-if where we did not before.
Use decltype to stop the instantiation of class templates
by making them depend on a parameter type.