yaml: explore the idea of passing configuration info in... config
Let's generate the config file, so that it also contains information
that currently we get from the (configure-generated) config.hh file.

Note that the runtime compilation by dyn:: defaults to using vcsn
compile which uses config.py which, so far, is unrelated to
config.yaml.  But the C++ code (export VCSN_NO_PYTHON to use it) does
work.

Also, we now have pkgdata in several places: srcdir and builddir.  So
we really need a data path, not just a data dir.

This shows that we abolutely _need_ an access to the config via
Python.

Issue #94.

* share/vcsn/config.yaml: Rename as...
* share/vcsn/config.in.yaml: this.
* configure.ac: Instantiate it.
* share/vcsn/local.mk: Adjust.
* tests/local.mk: Find it.

* lib/vcsn/misc/yaml.cc: Look for the config file is a list of
directories, not just a single place.
* lib/vcsn/dyn/translate.cc (config): New.
Use it in place of XGETENV and misc/config.hh.