automaton: a less-than comparison
There is one big change here: now the comparison between two automata
is no longer performed on the pretty-printing of both automata, but on
them being the same underlying set of transitions.

So in particular `0 -> 1 a` and `1 -> 0 a` used to be two different
named automata (when input this way), they are now equal, because in
reality, once the state name stripped, both are `0 -> 1 a`.

This comparison is of course much faster than the previous one.  It is
not clear, however, if the Vcsn platform really benefits from it in
any way.

    In [1]: import vcsn
    In [2]: a1 = vcsn.context('lal(abc), z').de_bruijn(100)
    In [3]: a2 = vcsn.context('lal(abc), z').de_bruijn(101)
    In [4]: timeit a1.format('daut') == a2.format('daut')
    1000 loops, best of 3: 1.41 ms per loop
    In [5]: timeit a1 == a2
    100000 loops, best of 3: 9.17 µs per loop

This is especially true for large automata that differs in the first
few transitions:

    In [6]: a1 = vcsn.context('lal(abc), b').random_automaton(1000)
    In [7]: a2 = vcsn.context('lal(abc), b').random_automaton(1000)
    In [8]: timeit a1.format('daut') == a2.format('daut')
    1 loop, best of 3: 746 ms per loop
    In [9]: timeit a1 == a2
    1000000 loops, best of 3: 2.4 µs per loop

Note that currently, because we compare only transitions, "unused"
states do not participate.  This is certainly a problem that should be
addressed.

Besides, it considers that automata with equal sets of transitions
but different *lists* of transitions to be different.  This is too
coarse and many tests no longer pass.  We cannot yet expose this
test to users.

* vcsn/misc/algorithm.hh (lexicographical_compare): New.
* vcsn/algos/less-than-automaton.hh: New.
* vcsn/dyn/algos.hh (less_than): New.
* vcsn/local.mk: Adjust.
* python/vcsn_cxx.cc: Bind it.