build: depend on the runtime version of Flex
Currently configure registers the installed version of Flex, and this
information is used in flex-lexer.hh to adapt to two APIs: before 2.6
and from 2.6.  Unfortunately, we ship our Flex-generated files, so it
may well happen that we ship files from Flex 2.6, and the user
compiles them as if they were generated by Flex 2.5.35.

Stop looking for the version at configure time, extract the version
information in our Flex wrapper.

Fixes #153.
Reported by Victor Miller.

* build-aux/m4/flex-version.m4: Remove.
* configure.ac: Don't call it.

* build-aux/bin/flex++.in: Generate scan-version.hh.
* lib/vcsn/dot/local.mk, lib/vcsn/dot/scan.hh,
* lib/vcsn/rat/local.mk, lib/vcsn/rat/scan.hh:
Use the generated scan-version.hh.
Since we now generate several files, we need to use the
same timestamp trick as with Bison.
While at it, generate the output in the srcdir, instead
of the build tree.

* vcsn/misc/flex-lexer.hh: Don't depend on config.hh.