docker: more improvements
* Dockerfile: Sort from least frequently updated to most frequently
updated, so that we preserve the interest of previous snapshots.
For the same reasons, split the installation of vcsn from that of
OpenFst.
Use multi-asignments with ENV.
Be sure to set ccache when we are `vcsn`, not `root`.  I did check
that it did not work before, and now it works.
Remove useless call to vcsn, call jupyter directly.
* debuild/Dockerfile: Likewise.