build: beware of broken YAML-cpp
The version 0.5.2 of Yaml-cpp is broken
(https://github.com/jbeder/yaml-cpp/pull/303#issuecomment-264647490).
Unfortunately Debian is *&@# slow at upgrading the unstable version to
0.5.3 (https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=847099).

Fortunately, Debian stable uses 0.5.1 which seems to work.

The broken feature is YAML::Node::remove.  Don't use it if configure
finds it is broken.  This means that we should not merge configuration
files, which should not happen with the test suite as of today.

See #94.

* configure.ac: Check if YAML::Node::remove works.
* lib/vcsn/misc/yaml.cc: Adjust accordingly.