config: generate a git-based version file
* share/vcsn/local.mk (share/vcsn/version.yaml): New.
Ship it, and create it at each run of `make`.
* vcsn/misc/yaml.hh (load_home_config): Remove.
* lib/vcsn/misc/yaml.cc: Load the version file.

* build-aux/bin/move-if-change: Do not use mv at all when not needed.
Without this change, distcheck fails, because the tarball is made with
a good version.yaml, it is extracted in readonly mode, and then we
generate a version.yaml.tmp and "mv" it as version.yaml.  This fails,
as the latter is readonly.
We should not mv at all in that case.