vcsn-compile: restore laziness
I broke the laziness in many ways in c54de952535d91c298ee6ef05b0d60215b41f0d2.

* libexec/vcsn-compile (target): Rename as...
(output): this, as if we were to introduce -o, --output, it would
exactly that.
(deps): New.   Use it.  Fixes a first bug, where shared libs created
a foo.d but we were loading foo.so.d.
(need_compiling): When reading a deps file, be ready to accept
some characters escaped (e.g., the spaces, and we do have spaces in
the plugin names), but also colons when they come by two (`::`) as
in `std::string`.
It should be noted that this is not supported by Make, but we don't
care (so far): we don't use make.