determinize/is-deterministic: make them instantiable for non-free labelsets
In vcsn:: we do have static_asserts: we do not want the user to use
determinize on invalid types.  But to set up dyn::, we want to compile
some algorithms ahead.

Our mega-context plugins have become a real problem: they are too big.
We want to change the way we instantiate algorithms in contexts, in
which case algorithms which cannot be instantiated at all in some
contexts are troublesome as they must be treated especially.  And
anyway the user gets nasty error messages:

    ----> 1 a.is_deterministic()

    RuntimeError: is_deterministic: requires free labelset
    is_deterministic: requires free labelset
      failed signature:
        mutable_automaton<context<nullableset<letterset<char_letters>>, b>>
      available versions:
        mutable_automaton<context<letterset<char_letters>, b>>
        mutable_automaton<context<letterset<char_letters>, q>>
        mutable_automaton<context<letterset<char_letters>, z>>
      failed command:
        vcsn compile -shared 'plugins/algos/is_deterministic/mutable_automaton<context<nullableset<letterset<char_letters>>, b>>.cc'

So let's handle this in the bridges: they raise when we use these
functions, so the user gets immediately a shorter error message:

    RuntimeError: is_deterministic: requires free labelset

and we can instantiate is-deterministic like any other algorithm.

* vcsn/algos/determinize.hh, vcsn/algos/is-deterministic.hh:
Make sure, in the bridges, that the labelsets are free.