demangle: don't write patterns that shadow themselves
Our patterns used to rewrite `(vcsn::)?(nullable)<({param})>` as
`\1<\2>`.  As a result `nullable<vcsn::nullable<foo>>` did not strip
the inner `vcsn::`, because the outer match (with the first `nullable`
did not change the string, and `re.sub` found nothing _after it_ to
replace.

So we must avoid useless rewritings.

* python/vcsn_tools/demangle.py: Fix a few patterns.
* tests/python/demangle.py: Add more test cases.