tupleset: fix generators
The current implementation is completely broken in the case of
tuplesets.  It's a very delicate exercise where we use
boost::range::join to prepend `\e` to the generators, then `cross` to
make a Cartesian product, and then `filter` to filter our `\e|\e` and
so forth.

The lifetime management became too hard to track in that case, and
ASAN was crazy on this piece of code.

It was much easier to completely change the approach.  Now in addition
to `generators`, labelsets provide a `pregenerators` function which
lists `\e` and then the generators if `has_one`, otherwise it's just
the same as `generators`.  Now `cross` features a `skip_first`
function, used by tupleset when `has_one` to skip `\e|\e`.

* vcsn/alphabets/setalpha.hh (generators, pregenerators): New.
* vcsn/labelset/genset-labelset.hh: Forward them.
* vcsn/labelset/labelset.hh, vcsn/labelset/letterset.hh,
* vcsn/labelset/nullableset.hh, vcsn/labelset/tupleset.hh: Implement them.
* vcsn/misc/cross.hh (skip_first): New.