build: remove useless uses of AC_DEFINE
Now that we have config.yaml, there is data that we no longer need in
config.hh.  Which can result in gratuitious recompilation if these
values are changed.

There are still things such as `PACKAGE_NAME` that we don't need, but
I don't know how to get rid of this (without having to wrap
autoheader, which I'd like to avoid :-).

Fixes #182.

* configure.ac: Don't define what we now get via config.yaml.