tests: improve notebook checkings
The handling of the widget changed in the newest versions of IPython.
Now ipynbdoctest sees some output, but it is not stored in the
notebook.  To handle this, we have to compare the reference dictionary
witht the effective dictionary side-by-side, which we did not do
before.

Take this as an opportunity to clean up various issues.

* tests/bin/test.py: Expose the counts of tests, passes etc.
* tests/bin/ipynbdoctest.py: Use them instead of duplicating them.
(compare_outputs): Rewrite as...
(check_outputs): this.
(run_cell): Rename i and n to n and i, as it is more logical.
Improve the display.
Remove redundant display (we used to display the cell numbers twice).

fixup! tests: improve notebook checkings