contexts: clean up the implementation of num_tapes
For some reason, the current implementation of num_tapes is not as
constexpr as I want it to be: there's a static_if which cannot use it.

So implement it as a traits instead.  While at it, clean up the
is_multitape traits, and split its implementation in the corresponding
file.

While at it, move it to its own file.