polynomialset: fix the elimination of zero labels
Because of an incorrect decltype, label_is_zero never worked!  As a
result, a polynomial of expressions such as `<1>\z` could exist,
instead of being reduced to `0`.

There will be tests in a forthcoming commit, via expansions.

* vcsn/weightset/polynomialset.hh (label_is_zero): Fix.
While at it, modernize: use detail::detect and detail::static_if.
This revealed that nullableset had (broken) support for
is_zero.  It was removed in the previous commit.

* vcsn/labelset/tupleset.hh (zero): Don't participate in overload if
not applicable.

* doc/notebooks/expression.automaton.ipynb: Update.