expansions: allow to experiment with non-denormalized expansions
This required that I disable asserts.  This is not sane, but I don't
have time ATM to study this, and I have fixed bugs in following
commits.  I will check this again at the end.

Besides, the old code is meant to be used with VCSN_DENORM.

* vcsn/core/rat/expansionset.hh: Try the support of non-denormalized
expansions with the new approach.