python: use timeouts on our subprocesses
We already saw our pipe of processes to convert dot to SVG
deadlocking.  We also currently have a problem on the BF where
ipnbdoctest does not terminate in some cases, which, after all,
_might_ come from the same issue.

So let's avoid these pipes and rather execute the processes in
sequence, and with a timeout.

* python/vcsn/dot.py (timeout): New.
Use it.
(_dot_to_svg): Don't play with three-process pipes.