rat: print aaaa the same way for lal and lan
For some reason, the treatment of lal is different from lan (and law).
This introduces gratuitous differences when applying Sarasvati
Moutoucomarapoulé's fusion of lal with lan.

Maybe the special case was meant to distinguish the case of wordset
where `aaaaa` is actually a single five-letter label, not a product of
`a`.  But in that case anyway `aaaaa` is different from `a{5}`, the
latter begin a product of single-letter labels.

Related to #23 and #24, and b9820c10f86096eb2e50efb234b1b5d457c29d2b.

* vcsn/core/rat/printer.hh, vcsn/core/rat/printer.hxx:
Don't special case lal.
* tests/python/expression.py, tests/rat/common.rat, tests/rat/q.rat:
Check it.