expansions: do not denormalize for conjunction
There is no need to denormalize, the code is not really harder to
write, but the result is really bad when we denormalize.  The reason
is the same compared to what we did with compose: a constant weight is
really as if it were attached to $ as first, not \e.  So for instance
the conjunction of the expansions of `(\e+a)` and `aa` gave two very
different results in lal (⟨0⟩) vs. lan (ε ⊙ [ε & aa]).

The key difference is that \e . [<k>1] matches a . [P]
whereas obviously          <k>         does not.

Denormalizing does simplify the code, but ruins the result.  To get
simple code, we should get rid of the constant code by attaching it to
the endmarker, `$`.

This will also help integrating Sarasvati Moutoucomarapoulé's fusion
of lal with lan, removing useless differences.

* vcsn/core/rat/expansionset.hh: Do not denormalize conjunctions.