random-label: new algorithm
The conversion from lal to lan introduced a problem: we are no longer
able to generate letter-labels that do not include the empty-word, so
we are no longer able to generate expressions that do not include
`\e`: it is "inherited" from the labelset.

To address this, we need to be able to control the generation of
random labels.  This is already what we do for expressions and
weights.  Besides, it appears that vcsn::random_label was not bound in
dyn.

* lib/vcsn/dyn/context-printer.hh, lib/vcsn/dyn/context-printer.cc
(headers_mid_): New level.
We need tuple to see the labelsets to be able to support random_label,
and we need algorithms to see tuple first.
This is really messy :(

* vcsn/labelset/labelset.hh (random_label): Remove, it should not be
needed.

* vcsn/labelset/letterset.hh, vcsn/labelset/oneset.hh,
* vcsn/labelset/tupleset.hh, vcsn/labelset/wordset.hh:
(random_label): Adjust to receive a param string.

* vcsn/dyn/algos.hh, python/vcsn_cxx.cc: Bind it.

* tests/python/random-label.py: Check it.