fst: support fstinfo
I wanted to know where OpenFST stands on is-deterministic when there
are spontaneous transitions.  The answer is: Open FST does not care
about them wrt determinism: the `one` label is treated like any other
label.

This is something I was considering for Vcsn too...  But I'm uneasy
with this, so I think I will rather introduce a new property P, and
define deterministic as having P and no spontaneous transitions.

See also #196.

* python/vcsn/automaton.py (fstinfo): New.
* tests/python/efsm.py: Check.