info: provide a 'strict' mode, which evaluates unknown properties
Currently we reported "N/A" for is-deterministic of any automaton.
That was not related to LAN/LAL, but rather to the fact that info does
not evaluate unknown properties, and unknown properties are displayed
as "N/A".

Reduces the number of failures to 5 (from 8).

* vcsn/algos/info.hh (info): Add a `strict` argument to decide whether
we should evaluate unknown properties.
Being able to not evaluate them is useful to check the property cache.