rat: convert the expression after parsing, if needed
Sometimes we get failures in random_expression for lat.  The reason is
that we generate a \e, and `exp('\e')` fails in lat.  There is code in
the parser to convert to multiband if needed, but it was not generic
enough.  And we actually already have solved this problem with `conv`
routines which use partial_identity to move from single-tape to
multitape.

However, `conv` is not accessible in dyn, but `copy` is.  But it is
very expensive on large expressions such as the ones used in vcsn
score.  So let's avoid this cost when it is not needed.  Unfortunately
we have no means to compare two contexts, so... we print their snames
and compare them.

* lib/vcsn/rat/parse.yy: here.