random-expression: fight harder for `length`
Currently the operators are picked independently of the current
length, so it was perfectly possible to draw a `\e` with length=10,
which is a sure means not to reach 10.  This is a problem for the
tests of random_expression itself as some tests do check that all the
demanded operators appear in a sufficiently long expression, but we
could very well (and it did happen too often) draw `\e`, which fails
the test, obviously.

Besides, labels were not treated as \e and \z (that's why we could
draw `\e`, but not `a` for instance).

Clean all this up.

There are many opportunities to speed this up:
- a random_label to save the parsing of the specs, like for weights
- caching the weights for nullary_ops_, unary_ops_, etc.

* vcsn/misc/random.hh (discrete_distribution::select): Remove.
(operator()): New overload which simplifies the caller (but is less
efficient, see above).

* vcsn/algos/random-expression.hh (nullary_exp_): New, to treat
labels like `\e` and `\z`.
(nullary_ops_, unary_ops_, binary_or_unary_ops_): New, to target
length with more certainty.
Support the "default" parameters.
* vcsn/dyn/algos.hh, python/vcsn_cxx.cc: Use that default.
* tests/python/factory.py: Update.

* doc/notebooks/context.random_expression.ipynb: Update.