score: prefer passing a environment to a bench than running a setup
The setup passed to timeit is run for each iteration.  That's very
costly (some cases are run 100,000 times), especially if we even move
the `import vcsn` into the setup: 8:50 to run `vcsn score` instead of
7:30.

However, evaluating the setup in the global context is unsafe,
especially if we want to run benches in parallel.  Fortunately, it is
possible to passe to timeit an environment (aka namespace), and `exec`
allows to specify the environment in which to run the commands.

Once put together, `vcsn score` runs in about 7.15.  However, at first
sight the result appear much better than before.  We might have
changed the meaning of our benches.

* libexec/vcsn-score: here.