dyn: simplify the type ast
I would like to have something more extensible, and extensible at
runtime: I'd like people to be able to design plugins for weightsets
or labelsets.

This requires that a simpler design.  Actually, Json seems like a nice
shot.  For a start, let's try to have an AST which is less typed, more
Json like.

* lib/vcsn/dyn: fuse letterset and wordset into genlabelset.