automaton: don't display context openness
Fixes #202.

We have a number of issues wrt whether the context of an automaton is
open or not (see #201 for instance).  This should use specific API to
test whether its context is open.

But it hardly makes sense to print that when displaying an automaton.

This causes problems in the test suite, see for instance
tests/python/complete.py.

* vcsn/misc/format.hh (show_open): New.
* vcsn/algos/daut.hh, vcsn/algos/dot.hh: Use it.