rat: let the parser display when it's performing conversion
We may have to pay for conversion (which can be costly, see
https://gitlab.lrde.epita.fr/vcsn/vcsn/issues/205) although the
expression does not need it.  That happens when for instance the
"input" context is empty and open, and the one we obtained contain
letters...  That's stupid, we should not pay for copy/conversion in
that case, but at least we need to be able to see it.