rat: let copy avoid useless copies
Somewhat related to #205: `copy does not preserve physical sharing`.

Currently the expression parser, when it has finished parsing,
compares the input context with the context of the produced
expression, and if they differ, it calls `copy` to convert the
expression to the input context.

This is because when we parse `a` in `lal<lal>`, we actually get `a`
as `lal` and then we copy-convert it to `a` in `lat<lal>`.

The comparison between the input context and the context of the result
is done on their snames (because so far we have nothing else to
compare two contexts).  This is fragile, and is actually broken by the
forthcoming change that shows openness of contexts in their snames.
So if we parse `a` with `[...] -> b`, we actually get `a` with `[a] ->
b` and so we copy this `a`.  This shows in particular when parsing
`a{5}` because as of today copy does not preserve physical sharing, so
the parser breaks the nice sharing of `a` in `a{5}` to end with five
copies of `a` (this is issue #205).

It turns out that since expressions are immutable, `copy` should just
as well _not_ copy when it's useless.  This simplifies the parser, is
faster (no string comparisons of snames), and is actually saving from
more useless conversions (since we do not need to convert `a{5}` when
going from `[a] -> b` to `[ab] -> b` or any other sort of useless case
where the context types are actually the same).

Tests are needed...

* vcsn/core/rat/copy.hh: Avoid useless copies.
* lib/vcsn/rat/parse.yy: Always copy/convert: let it do its job.
* vcsn/ctx/instantiate.hh: Be sure to instantiate it by default,
now that it's used when parsing any expression.
Adjust dependencies.