is-ambiguous: refactor to exhibit an `on_diagonal` function
In e9ce3ed7425e8b62f8ba440f7a29cc9a825f9f89, Sarasvati
Moutoucomarapoulé fixed an issue caused by the removal of lan: now
that all automata can have spontaneous transitions, all our
conjunction insplit the second argument, hence when we compared the
states on the diagonal, we must check the pre-insplit state.

However we also need that precaution to be taken in the case of
is_cycle_ambiguous, so make her approach public.  Will be used in the
next commit.

The witness highly depends on the order in which states are checked.
Here, because we no longer iterate on the origin, we changed the
order, hence a test had to be adjusted.

There appears to be no visible impact for vcsn score.

See also #206 and #207.