is-ambiguous: require a free automaton
Fixes #207.

Of course the automaton must be letterized for the algo to work.  But
it also reequires to be proper because we check if we stay on the
diagonal of `aut & aut`, and our implementation of conjunction with
spontaneous transitions does not stay on the diagonal: we first follow
the spontaneous transition of the first automaton, then we follow that
of the second.

Maybe there is a way to cope with these sequences of spontaneous
transitions that "morally" stay on the diagonal.  Maybe there's a way
to stay on the diagonal in our conjunction (that what the papers from
Mohri claim doing, however if we look at the result of an fstintersect
it seems to do two transitions too).

But in any case, as of today, this does not work.

* vcsn/algos/is-ambiguous.hh: Require free automata.

* doc/notebooks/automaton.is_ambiguous.ipynb,
* doc/notebooks/automaton.is_cycle_ambiguous.ipynb,
* tests/python/is-ambiguous.py:
Adjust.