dyn: ctx: make sure not to glue multiple errors together
Currently, because we want to avoid the spurious final `\n` in our
exceptions, we do not append the final `\n` to the error messages.
But there can be several errors, in which case the separating `\n` is
missing.

Always add it, but remove the trailing `\n` when raising an error.