rat: printer: print `\e + E` as `E?`
This is more difficult than appears at first sight: we have to emulate
a fictional unary postfix operator `?`.  To this end, we must have
`\e + E` present itself to its parent as a unary operator, not as a
sum (other we would print `(a?)*` instead of `a?*`), and to its
children it must still behave as a sum.

We also need print_add_ to be ready to skip the first member, i.e., to
print `E...` in `\e + E...`.