doc: update table style in NEWS
Unfortunately the syntax of the tables supported by GFM (Gitlab's
Markdown) is poor.  It does support <table>, but then the text file is
hardly readable.  We use the richer syntax supported by pandoc (in
particular is is possible to have multiline cells, something that GFM
does not support), but indent the tables so that Gitlab renders them
as plain text.

We could two NEWS: one for the human reader, and one for gitlab, but
gitlab does not allow us to specify the path to its news file, so
there would be two news files in the top-level.  That seems wrong.