Sms2fr: use xz
Now that we support compressed efsm, let's compress the efsm
files of sms2fr.  Not only will this save space, it should also
help with sites that claim that Vcsn is 98.3% of shell-scripts...

The results with Xz are impressive: graphemic.efsm is reduced to
5% of its original size, and syntactic.efsm 12%.

    -rw-r--r--   1 akim  staff  93099091 21 mar 23:46 graphemic.efsm
    -rw-r--r--   1 akim  staff  11713613 21 mar 19:05 graphemic.efsm.bz2
    -rw-r--r--   1 akim  staff   5502716 21 mar 23:22 graphemic.efsm.xz
    -rw-r--r--   1 akim  staff  62289670 21 mar 23:46 syntactic.efsm
    -rw-r--r--   1 akim  staff  13055960 13 mar 19:44 syntactic.efsm.bz2
    -rw-r--r--   1 akim  staff   7916576 21 mar 23:24 syntactic.efsm.xz

And Vcsn is now 68.8% C++, 20.6% Python and 3.5% shell-script!

While at it:
- use int not char for options
- don't use references to keep the automata
- don't force the format to efsm: let vcsn guess the format