edit-automaton: don't insert transitions with the special label
We parse in `lan, q`, we parse `<1/2>` as `<1/2>$`, not as `<1/2>\e`.
This is to support pre- and post-transition labels (e.g., `$ 0 <1/2>`).
Unfortunately it's easy, in particular with lao, to forget to write
`\e`, and I have been bitten: I tried to get the weight-series

    context = lao, expressionset<lal, q>
    $ -> 0
    0 -> 1 <a>
    1 -> 1 <b>
    1 -> 2 <c>
    2 -> $

But I got `<a+c>`.  The culprit was actually proper/is_proper: both
consider that automaton to be proper.  And the real culprit is
actually the user: there are indeed no spontaneous transitions in this
automaton: they are labeled with `$` because we did not use `\e`.
This error is too easy, so be sure to be kind to the user, and map the
special label to one when building the automaton.

It turns out that the test suite itself contained such an error: a
proper automaton still had a pseudo-spontaneous transition.

* vcsn/algos/edit-automaton.hh: Map $ to \e when adding an inner
transition.
* tests/python/proper.dir/lao-r.out.gv: Fix.