dyn: print in utf8 by default
We are not ready yet to parse labels in utf8, because we parse
directly by hand on top of the stream, which provides us only with a
backtracking/lookahead for 1 byte.  This is enough for `<`, but not
for the UTF-8 equivalent which is several bytes long.

So currently, don't print dot/daut with labels in UTF-8.

Still, let C++ target utf-8 by default when possible, and have Python
just use the default rather that implementing another on top.

* lib/vcsn/algos/print.cc: here.

* python/vcsn/context.py, python/vcsn/expansion.py,
* python/vcsn/expression.py, python/vcsn/ipython.py,
* python/vcsn/label.py, python/vcsn/polynomial.py
* python/vcsn/weight.py:
Rely on this default.

Adjust dependencies.