core/tuple-automaton: declare `indices` at namespace scope
In C++14, static constexpr expressions that are odr-used must be
declared in the compilation unit at namespace scope even if they are
defined and assigned to in the class.

The `indices` static constexpr attribute of tuple_automaton_impl was
missing that namespace-level declaration, some compiler were fine with
it, but clang 8 isn't, unless compiling with -O1 or higher.

Fix #216