core/tuple-automaton: make `indices` inline
In C++14, static constexpr expressions that are odr-used must be
declared in the compilation unit at namespace scope even if they are
defined and assigned to in the class.

C++17 gets rid of that requirement by introducing the concept of inline
constexpr and making constexpr inline by default. Some compilers allow
the `inline` keyword on constexpr in C++14 as a C++17 extension.

Fix #216